AuthServer

+++ AutServer.Server/Abstract/IPersonService.cs
@@ -5,10 +5,10 @@ namespace AutServer.Server.Abstract
 {
     public interface IPersonService
     {
-        Task<int> Add(Person entity, IFormFile formFile);
-        Task<int> Delete(Person entity);
-        Task<int> Update(Person entity, IFormFile formFile);
+        Task<long> Add(Person entity, IFormFile formFile);
+        Task<long> Delete(Person entity);
+        Task<long> Update(Person entity, IFormFile formFile);
         Task<List<Person>> GetAll();
-        Task<Person> GetById(int Id);
+        Task<Person> GetById(long Id);
     }
 }
